refactor components into files
